requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv    
pymupdf  
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
streamlit
langchain_groq
pandas
openai
duckduckgo_search
validators==0.28.1
youtube_transcript_api
pytube
unstructured
numexpr
SQLAlchemy
mysql-connector-python
gradio